Added updated recipe DB
Added 7 more recipes, as well as a "UserAdded" column for the recipe table, we will be using 'n' for if we added them ourselves, and 'y' if the user added the recipe. For debugging purposes, I have updated the java code to see the data in the recipes table after this update. It is as followed and can replace the code to see data in recipes table that was there before:

            ////////////////////////////////////
            //To see all data in recipes table//
            ////////////////////////////////////
            String queryRecipeData = "select * from recipes";
            System.out.println("\nData from recipes table");
            System.out.printf("%5s%20s%20s%50s%35s%n", "ID", "NAME", "serving size" , " link", "userAdded");
            rset = statement.executeQuery(queryRecipeData);

            while(rset.next()) {   // process each row
                int idRecipe = rset.getInt("idRecipes");
                String recipeName = rset.getString("name");
                int amountFed = rset.getInt("amountFed");
                String urlLink = rset.getString("link");
                String userAdded = rset.getString("userAdded");
                System.out.printf("%5d%30s%5d%80s%5s%n",idRecipe, recipeName, amountFed, urlLink, userAdded);
            }