Created Recipe class and everything needed for recipe objects
Created a recipe class we can choose to implement/edit if we keep, as well as all the logic needed to create efficient recipe classes that shows their own name, serving size, web address, and ingredients, as well as how to output each of those. the Recipe class I was using is at the bottom below the commented code to see the tables.